requirements.txt
# Core dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pydantic-settings>=2.6.0

# LangChain and LangGraph
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.0
langchain-core>=0.3.0

# Web search
tavily-python>=0.5.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.27.0
sse-starlette>=2.1.0

# Optional testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
